21.3/1

+++ src/main/java/com/crud/tasks/controller/TaskController.java
@@ -32,8 +32,12 @@ public class TaskController {
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "deleteTask")
-    public void deleteTask(@RequestParam Long taskId) {
-        service.deleteTask(taskId);
+    public void deleteTask(@RequestParam Long taskId) throws TaskNotFoundException {
+        if (service.getTask(taskId).isPresent()) {
+            service.deleteTask(taskId);
+        } else {
+            throw new TaskNotFoundException();
+        }
     }
 
     @RequestMapping(method = RequestMethod.PUT, value = "updateTask")
